requirements.txt
fastapi
uvicorn
mysql-connector-python
python-jose
python-multipart
passlib[bcrypt]
bcrypt==3.2.0